refactor package naming: Rename more_options to moreoptions, training_entities to trainingentities, and profileSelector to profileselector
