+++ JetFit/app/src/main/java/com/google/jetfit/presentation/App.kt
@@ -15,13 +15,13 @@ import androidx.navigation.navArgument
 import com.google.jetfit.presentation.screens.Screens
 import com.google.jetfit.presentation.screens.favorites.FavoritesScreen
 import com.google.jetfit.presentation.screens.home.HomeScreen
-import com.google.jetfit.presentation.screens.more_options.MoreOptionsScreen
+import com.google.jetfit.presentation.screens.moreoptions.MoreOptionsScreen
 import com.google.jetfit.presentation.screens.player.audio.AudioPlayerScreen
 import com.google.jetfit.presentation.screens.player.video.VideoPlayerScreen
-import com.google.jetfit.presentation.screens.profileSelector.ProfileSelectorScreen
+import com.google.jetfit.presentation.screens.profileselector.ProfileSelectorScreen
 import com.google.jetfit.presentation.screens.subscription.SubscriptionScreen
 import com.google.jetfit.presentation.screens.training.TrainingScreen
-import com.google.jetfit.presentation.screens.training.training_entities.TrainingEntityScreen
+import com.google.jetfit.presentation.screens.training.trainingentities.TrainingEntityScreen
 import com.google.jetfit.presentation.utils.navigateTo
 import com.google.jetfit.presentation.utils.navigationDrawerGraph
