+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/dashboard/DashboardScreen.kt
@@ -16,14 +16,14 @@ import androidx.navigation.compose.composable
 import com.google.jetfit.presentation.screens.Screens
 import com.google.jetfit.presentation.screens.favorites.FavoritesScreen
 import com.google.jetfit.presentation.screens.home.HomeScreen
-import com.google.jetfit.presentation.screens.more_options.MoreOptionsScreen
+import com.google.jetfit.presentation.screens.moreoptions.MoreOptionsScreen
 import com.google.jetfit.presentation.screens.player.audio.AudioPlayerScreen
 import com.google.jetfit.presentation.screens.player.video.VideoPlayerScreen
 import com.google.jetfit.presentation.screens.search.SearchScreen
 import com.google.jetfit.presentation.screens.settings.SettingsScreen
 import com.google.jetfit.presentation.screens.subscription.SubscriptionScreen
 import com.google.jetfit.presentation.screens.training.TrainingScreen
-import com.google.jetfit.presentation.screens.training.training_entities.TrainingEntityScreen
+import com.google.jetfit.presentation.screens.training.trainingentities.TrainingEntityScreen
 
 
 @Composable
