+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/moreoptions/composable/BackRowSchema.kt
@@ -1,4 +1,4 @@
-package com.google.jetfit.presentation.screens.more_options.composable
+package com.google.jetfit.presentation.screens.moreoptions.composable
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
