+++ JetFit/app/src/main/java/com/google/jetfit/presentation/screens/training/trainingentities/composables/RoundedGradientImage.kt
@@ -1,4 +1,4 @@
-package com.google.jetfit.presentation.screens.training.training_entities.composables
+package com.google.jetfit.presentation.screens.training.trainingentities.composables
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
